Implementing a structure diagram using mermaid

+++ src/cpu/bus.py
@@ -9,14 +9,14 @@ class Bus:
         self.wram = Memory()
         self.vram = None
 
-    def write(self, addr: np.uint8, value: np.uint8) -> None:
+    def write(self, addr: np.uint16, value: np.uint8) -> None:
         self.wram.write(addr, value)
     
-    def read(self, addr: np.uint8) -> np.uint8:
+    def read(self, addr: np.uint16) -> np.uint8:
         return self.wram.read(addr)
     
-    def write_u16(self, addr: np.uint8, value: np.uint16) -> None:
+    def write_u16(self, addr: np.uint16, value: np.uint16) -> None:
         self.wram.write_u16(addr, value)
     
-    def read_u16(self, addr: np.uint8) -> np.uint16:
+    def read_u16(self, addr: np.uint16) -> np.uint16:
         return self.wram.read_u16(addr)
